WebSockets and IPC changes after review
