+++ src/Nethermind/Nethermind.Api/IApiWithNetwork.cs
@@ -28,7 +28,7 @@ using Nethermind.Stats;
 using Nethermind.Synchronization;
 using Nethermind.Synchronization.ParallelSync;
 using Nethermind.Synchronization.Peers;
-using Nethermind.WebSockets;
+using Nethermind.Sockets;
 
 namespace Nethermind.Api
 {
