+++ src/Nethermind/Nethermind.Api/Nethermind.Api.csproj
@@ -13,7 +13,7 @@
       <ProjectReference Include="..\Nethermind.JsonRpc\Nethermind.JsonRpc.csproj" />
       <ProjectReference Include="..\Nethermind.Monitoring\Nethermind.Monitoring.csproj" />
       <ProjectReference Include="..\Nethermind.Network\Nethermind.Network.csproj" />
-      <ProjectReference Include="..\Nethermind.WebSockets\Nethermind.WebSockets.csproj" />
+      <ProjectReference Include="..\Nethermind.Sockets\Nethermind.Sockets.csproj" />
     </ItemGroup>
 
 </Project>
