+++ src/Nethermind/Nethermind.DataMarketplace.Infrastructure/NdmApi.cs
@@ -70,7 +70,7 @@ using Nethermind.Synchronization.Peers;
 using Nethermind.Trie.Pruning;
 using Nethermind.TxPool;
 using Nethermind.Wallet;
-using Nethermind.WebSockets;
+using Nethermind.Sockets;
 
 namespace Nethermind.DataMarketplace.Infrastructure
 {
