+++ src/Nethermind/Nethermind.DataMarketplace.WebSockets.Test/NdmWebSocketsConsumerChannelTests.cs
@@ -17,7 +17,7 @@
 using FluentAssertions;
 using Nethermind.Core.Crypto;
 using Nethermind.DataMarketplace.Channels;
-using Nethermind.WebSockets;
+using Nethermind.Sockets;
 using NSubstitute;
 using NUnit.Framework;
