+++ src/Nethermind/Nethermind.JsonRpc/Nethermind.JsonRpc.csproj
@@ -32,6 +32,6 @@
     <ProjectReference Include="..\Nethermind.Monitoring\Nethermind.Monitoring.csproj" />
     <ProjectReference Include="..\Nethermind.Network\Nethermind.Network.csproj" />
     <ProjectReference Include="..\Nethermind.Wallet\Nethermind.Wallet.csproj" />
-    <ProjectReference Include="..\Nethermind.WebSockets\Nethermind.WebSockets.csproj" />
+    <ProjectReference Include="..\Nethermind.Sockets\Nethermind.Sockets.csproj" />
   </ItemGroup>
 </Project>
