+++ src/Nethermind/Nethermind.Pipeline/Nethermind.Pipeline.csproj
@@ -1,9 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <ItemGroup>
-    <ProjectReference Include="..\Nethermind.WebSockets\Nethermind.WebSockets.csproj" />
+    <ProjectReference Include="..\Nethermind.Sockets\Nethermind.Sockets.csproj" />
   </ItemGroup>
-
   <ItemGroup>
     <PackageReference Include="System.IO.Abstractions" Version="13.2.9" />
   </ItemGroup>
