+++ src/Nethermind/Nethermind.Runner.Test/Ethereum/ContextWithMocks.cs
@@ -51,7 +51,7 @@ using Nethermind.Synchronization.Peers;
 using Nethermind.Trie.Pruning;
 using Nethermind.TxPool;
 using Nethermind.Wallet;
-using Nethermind.WebSockets;
+using Nethermind.Sockets;
 using NSubstitute;
 
 namespace Nethermind.Runner.Test.Ethereum
