+++ src/Nethermind/Nethermind.Runner/Ethereum/Api/NethermindApi.cs
@@ -62,7 +62,7 @@ using Nethermind.Synchronization.Peers;
 using Nethermind.Trie.Pruning;
 using Nethermind.TxPool;
 using Nethermind.Wallet;
-using Nethermind.WebSockets;
+using Nethermind.Sockets;
 
 namespace Nethermind.Runner.Ethereum.Api
 {
