+++ src/Nethermind/Nethermind.Runner/configs/calaveras.cfg
@@ -16,7 +16,6 @@
   },
   "JsonRpc": {
     "Enabled": false,
-    "IpcUnixDomainSocketPath": "C:\\temp\\nethIpc.soc",
     "TracerTimeout": 20000,
     "Host": "127.0.0.1",
     "Port": 8559
