+++ src/Nethermind/Nethermind.Sockets/ISocketHandler.cs
@@ -4,8 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Nethermind.WebSockets
+namespace Nethermind.Sockets
 {
+    /// <summary>
+    /// Interface that provides lower level operations (in comparison to <see cref="ISocketsClient"/>)
+    /// from a specific socket implementation like for example WebSockets, UnixDomainSockets or network sockets.
+    /// </summary>
     public interface ISocketHandler : IDisposable
     {
         Task SendRawAsync(string data);
