+++ src/Nethermind/Nethermind.Sockets/SocketsMessage.cs
@@ -14,7 +14,7 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
 
-namespace Nethermind.WebSockets
+namespace Nethermind.Sockets
 {
     public class SocketsMessage
     {
