+++ src/Nethermind/Nethermind.Sockets/WebSocketsManager.cs
@@ -14,16 +14,26 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
 
-using System;
-using System.Threading.Tasks;
+using System.Collections.Concurrent;
 
-namespace Nethermind.WebSockets
+namespace Nethermind.Sockets
 {
-    public interface ISocketsClient : IDisposable
+    public class WebSocketsManager : IWebSocketsManager
     {
-        string Id { get; }
-        string ClientName { get; }
-        Task ReceiveAsync();
-        Task SendAsync(SocketsMessage message);
+        private readonly ConcurrentDictionary<string, IWebSocketsModule> _modules = new();
+
+        private IWebSocketsModule _defaultModule = null;
+
+        public void AddModule(IWebSocketsModule module, bool isDefault = false)
+        {
+            _modules.TryAdd(module.Name, module);
+            
+            if (isDefault)
+            {
+                _defaultModule = module;
+            }
+        }
+
+        public IWebSocketsModule GetModule(string name) => _modules.TryGetValue(name, out var module) ? module : _defaultModule;
     }
 }
